Simple scripts to get the year to date stats for the athletes of an Strava club.
This 3 scripts help to get the year to date distance traveled and elevation gain
for a group of athletes. This is done by reading the public profile page of each
athlete and extracting the information and converting them into metric. Each
script can be run as independently in the command line or can be imported as a
module.

There are two parts of this scripts, both have some limitations.

- extract_group_members.py: extracts the members of a strava club. This
  scripts needs the saved HTML from the strava club website saved by the owner
  of the club.

- get_stats.py: For a list of Strava Athletes extract the Year to date from the
  public web page. This may fail depending on the users preference (Cycling vs
  Running) or if the user is a Strava Pro Athlete.

- get_ytd_club_stats.py: Calls the previous two scripts in sequence, has the
  same limitations and assumptions of the previous two scripts.