Fix csv header csv file header to lowercase to get round Excel warning.
Excel gives a warning indicating the file is in SYLK format. As reported
by https://support.microsoft.com/en-us/kb/215591
Excel assumes a csv file is in SYLK format if the first two characters of
the file start with uppercase ID. Use lowercase id as the header to remove
this warning.